Add frontend application (#48)
* Update build.gradle

- Java version set to 1.8
- Example code commented

* Update build.gradle

Gradle version changed 5.2.1 -> 6.1.1

* Add initial frontend

Add the initial frontend for our server, does not show any proper info yet but it is able to start up.

Run 'npm install' within the website folder and then 'npm start' to boot up the server.

Related to issue #24

* Add initial list component

Use material-ui list components to build a presentable list for the CI build list. No real data is shown yet, but the design is there. Idea is to have a minimalistic view showing the build information.

Color scheme used: https://coolors.co/06aed5-086788-f0c808-fff1d0-dd1c1a

Related to issue #24

* Rename website folder

Rename the website folder to frontend to keep the naming convention.

Related to issue #24

* Update .gitignore

Add node_modules to the .gitignore

* Add Storage class

Add initial format for Storage class, with a constructor implementation that creates a new local json file if none exists already.

Two functions will be available for use;
* Get
* Post

Which will work similar to most APIs where we are able to get and post into the json storage object.

Related to issue #17

* Add post function

Post function now works as intended and adds the storage object (with a maybe to change format) to the database.

Related to issue #17 and #25

* Add Get function

Get function should now work as intended and fetches a given object from the database depending on its id. If key was not found in object, then return an empty object.

Related to issue #17 and #25

* Add gradle dependencies

Add some new gradle dependencies that were used for Storage

* Integrate Build class into storage

Storage and API skeleton code are now working together. API is now able to return a list of Build objects that should be fetched from the frontend. This is done by performing a get request to <your local address>/ci/get and you will get a reply in the following format:

* jobID
  * status
  * commitSha
  * url
  * log

Add getAll function in Storage class so that all builds can be fetched at the same time for performance and ease of use reasons.

Related to issue #26

* Fix naming

Naming was broken after merge, should be fixed now.

* Add get call from frontend

Frontend is now able to fetch the database by using a proxy to bypass CORS. Try to integrate it without the proxy if possible.

Related to issue #24

* Make frontend dynamic

Frontend now fetches from the database and creates a list item for each one. It does not contain the exact information we want to show yet however, so it is not finished.

Added dotenv to store ngrok and proxy urls.

In order to run, add your ngrok url (not localhost) to the .env file and then start the frontend by running 'npm start'. Other than this, the ngrok server and backend server also have to be running.

Updated .gitignore to hide .env file which means you have to create your own in order to start the server. Ask me for more information about how to set it up.

This is related to issue #24 and #40

* Design and formatting

Add basic formatting and design for the list all builds view

* Add Single build info page

Now able to click on a specific build and it takes you to a page that gives you more information regarding it. You can then click the box once again to go back to the original page. Displays commit sha, job id, status, build info and test info as according to issue #25.

Routing was done by using react router.

* Update env to js file

It doesn't work to have a .env so we use env.js instead.

* Display build data

* Add build list url to readme

* Updates readme

Co-authored-by: Adam Björnberg <38786385+adbjo@users.noreply.github.com>
Co-authored-by: Johan Moritz <jmoritz@kth.se>