Adds unit-tests for Resource.java and Storage.java (#70)
* Add ResourceTest and additional StorageTest

Add test for:
* Fetching data using backend api
* Clearing database
* Fetching ALL data using getAll instead of just single instances

Add functionality for:
* Clearing database

Reasoning behind adding additional Storage tests here, as well as a clear function, was to properly be able to run the backend api tests. Database had to be able to be cleared before running these. API also uses getAll and those tests were missing in the StorageTests class so they were added quickly.

Related to issue #67 and #60 

* Adds unit-tests for POST request handling

- test1 tests if a webhook creation post request is handled.
- test2 tests if a push POST request is handled.
- Also removed unused imports in ResourceTest and old coment in Resource.java.

Related to issue #67 and #60 

Co-authored-by: Gabriel Gessle <sandshrew97@gmail.com>